[Puppi] changed most doubles to floats

+++ CommonTools/PileupAlgos/interface/PuppiCandidate.h
@@ -2,11 +2,11 @@
 #define CommonTools_PileupAlgos_PuppiCandidate_h
 
 struct PuppiCandidate {
-  double pt{0};
-  double eta{0};
-  double phi{0};
-  double m{0};
-  double rapidity{0};
+  float pt{0};
+  float eta{0};
+  float phi{0};
+  float m{0};
+  float rapidity{0};
   int id{0};
 };
