+++ CommonTools/PileupAlgos/src/PuppiContainer.cc
@@ -37,7 +37,7 @@ void PuppiContainer::initialize(const std::vector<RecoObj> &iRecoObjects) {
   fRawAlphas.resize(0);
   fAlphaMed.resize(0);
   fAlphaRMS.resize(0);
-  fNPV = 1.;
+  fNPV = 1;
   //Link to the RecoObjects
   fRecoParticles = &iRecoObjects;
   fPFParticles.reserve(iRecoObjects.size());
@@ -53,11 +53,11 @@ void PuppiContainer::initialize(const std::vector<RecoObj> &iRecoObjects) {
       pCand.phi = rParticle.phi;
       pCand.m = rParticle.m;
     } else {
-      pCand.pt = 0.;
-      pCand.eta = 99.;
-      pCand.rapidity = 99.;
-      pCand.phi = 0.;
-      pCand.m = 0.;
+      pCand.pt = 0.f;
+      pCand.eta = 99.f;
+      pCand.rapidity = 99.f;
+      pCand.phi = 0.f;
+      pCand.m = 0.f;
     }
 
     fPFParticles.push_back(pCand);
@@ -77,22 +77,22 @@ void PuppiContainer::initialize(const std::vector<RecoObj> &iRecoObjects) {
 
 PuppiContainer::~PuppiContainer() {}
 
-double PuppiContainer::goodVar(PuppiCandidate const &iPart,
+float PuppiContainer::goodVar(PuppiCandidate const &iPart,
                                std::vector<PuppiCandidate> const &iParts,
                                int iOpt,
-                               const double iRCone) {
+                               const float iRCone) {
   return var_within_R(iOpt, iParts, iPart, iRCone);
 }
 
-double PuppiContainer::var_within_R(int iId,
+float PuppiContainer::var_within_R(int iId,
                                     const vector<PuppiCandidate> &particles,
                                     const PuppiCandidate &centre,
-                                    const double R) {
+                                    const float R) {
   if (iId == -1)
-    return 1.;
+    return 1.f;
 
-  double const r2 = R * R;
-  double var = 0.;
+  float const r2 = R * R;
+  float var = 0.f;
 
   for (auto const &cand : particles) {
     if (std::abs(cand.rapidity - centre.rapidity) < R) {
@@ -141,11 +141,11 @@ void PuppiContainer::getRMSAvg(int iOpt,
     //Get the Puppi Sub Algo (given iteration)
     int pAlgo = fPuppiAlgo[pPupId].algoId(iOpt);
     bool pCharged = fPuppiAlgo[pPupId].isCharged(iOpt);
-    double pCone = fPuppiAlgo[pPupId].coneSize(iOpt);
+    float pCone = fPuppiAlgo[pPupId].coneSize(iOpt);
     // compute the Puppi metric:
     //  - calculate goodVar only for candidates that (1) will not be assigned a predefined weight (e.g 0, 1),
     //    or (2) are required for computations inside puppi-algos (see call to PuppiAlgo::add below)
-    double pVal = -1;
+    float pVal = -1;
     bool const getsDefaultWgtIfApplyCHS = iConstits[i0].id == 1 or iConstits[i0].id == 2;
     if (not((fApplyCHS and getsDefaultWgtIfApplyCHS) or iConstits[i0].id == 3) or
         (std::abs(iConstits[i0].eta) < fPuppiAlgo[pPupId].etaMaxExtrap() and getsDefaultWgtIfApplyCHS)) {
@@ -192,9 +192,9 @@ void PuppiContainer::getRawAlphas(int iOpt,
       //Get the Puppi Sub Algo (given iteration)
       int pAlgo = fPuppiAlgo[j0].algoId(iOpt);
       bool pCharged = fPuppiAlgo[j0].isCharged(iOpt);
-      double pCone = fPuppiAlgo[j0].coneSize(iOpt);
+      float pCone = fPuppiAlgo[j0].coneSize(iOpt);
       //Compute the Puppi Metric
-      double const pVal = goodVar(iConstits[i0], pCharged ? iChargedParticles : iParticles, pAlgo, pCone);
+      float const pVal = goodVar(iConstits[i0], pCharged ? iChargedParticles : iParticles, pAlgo, pCone);
       fRawAlphas.push_back(pVal);
       if (!edm::isFinite(pVal)) {
         LogDebug("NotFound") << "====> Value is Nan " << pVal << " == " << iConstits[i0].pt << " -- "
@@ -222,22 +222,22 @@ int PuppiContainer::getPuppiId(float iPt, float iEta) {
   //if(lId == -1) std::cerr << "Error : Full fiducial range is not defined " << std::endl;
   return lId;
 }
-double PuppiContainer::getChi2FromdZ(double iDZ) {
+float PuppiContainer::getChi2FromdZ(float iDZ) {
   //We need to obtain prob of PU + (1-Prob of LV)
   // Prob(LV) = Gaus(dZ,sigma) where sigma = 1.5mm  (its really more like 1mm)
-  //double lProbLV = ROOT::Math::normal_cdf_c(std::abs(iDZ),0.2)*2.; //*2 is to do it double sided
+  //float lProbLV = ROOT::Math::normal_cdf_c(std::abs(iDZ),0.2)*2.; //*2 is to do it double sided
   //Take iDZ to be corrected by sigma already
-  double lProbLV = ROOT::Math::normal_cdf_c(std::abs(iDZ), 1.) * 2.;  //*2 is to do it double sided
-  double lProbPU = 1 - lProbLV;
+  float lProbLV = ROOT::Math::normal_cdf_c(std::abs(iDZ), 1.) * 2.f;  //*2 is to do it double sided
+  float lProbPU = 1 - lProbLV;
   if (lProbPU <= 0)
     lProbPU = 1e-16;  //Quick Trick to through out infs
   if (lProbPU >= 0)
     lProbPU = 1 - 1e-16;  //Ditto
-  double lChi2PU = TMath::ChisquareQuantile(lProbPU, 1);
+  float lChi2PU = TMath::ChisquareQuantile(lProbPU, 1);
   lChi2PU *= lChi2PU;
   return lChi2PU;
 }
-std::vector<double> const &PuppiContainer::puppiWeights() {
+std::vector<float> const &PuppiContainer::puppiWeights() {
   int lNParticles = fRecoParticles->size();
 
   fWeights.clear();
@@ -257,12 +257,12 @@ std::vector<double> const &PuppiContainer::puppiWeights() {
   if (fPuppiDiagnostics)
     getRawAlphas(0, fPFParticles, fPFParticlesForVar, fPFParticlesForVarChargedPV);
 
-  std::vector<double> pVals;
+  std::vector<float> pVals;
   pVals.reserve(lNParticles);
   for (int i0 = 0; i0 < lNParticles; i0++) {
     //Refresh
     pVals.clear();
-    double pWeight = 1;
+    float pWeight = 1;
     //Get the Puppi Id and if ill defined move on
     const auto &rParticle = (*fRecoParticles)[i0];
     int pPupId = getPuppiId(rParticle.pt, rParticle.eta);
@@ -274,7 +274,7 @@ std::vector<double> const &PuppiContainer::puppiWeights() {
     }
 
     // fill the p-values
-    double pChi2 = 0;
+    float pChi2 = 0;
     if (fUseExp) {
       //Compute an Experimental Puppi Weight with delta Z info (very simple example)
       pChi2 = getChi2FromdZ(rParticle.dZ);
@@ -309,11 +309,11 @@ std::vector<double> const &PuppiContainer::puppiWeights() {
     // Protect high pT photons (important for gamma to hadronic recoil balance)
     if (fPtMaxPhotons > 0 && rParticle.pdgId == 22 && std::abs(fPFParticles[i0].eta) < fEtaMaxPhotons &&
         fPFParticles[i0].pt > fPtMaxPhotons)
-      pWeight = 1.;
+      pWeight = 1.f;
     // Protect high pT neutrals
     else if ((fPtMaxNeutrals > 0) && (rParticle.id == 0))
       pWeight = std::clamp(
-          (fPFParticles[i0].pt - fPtMaxNeutralsStartSlope) / (fPtMaxNeutrals - fPtMaxNeutralsStartSlope), pWeight, 1.);
+          (fPFParticles[i0].pt - fPtMaxNeutralsStartSlope) / (fPtMaxNeutrals - fPtMaxNeutralsStartSlope), pWeight, 1.f);
     if (pWeight < fPuppiWeightCut)
       pWeight = 0;  //==> Elminate the low Weight stuff
     if (fInvert)
@@ -326,7 +326,7 @@ std::vector<double> const &PuppiContainer::puppiWeights() {
     //Now get rid of the thrown out weights for the particle collection
 
     // leave these lines in, in case want to move eventually to having no 1-to-1 correspondence between puppi and pf cands
-    // if( std::abs(pWeight) < std::numeric_limits<double>::denorm_min() ) continue; // this line seems not to work like it's supposed to...
+    // if( std::abs(pWeight) < std::numeric_limits<float>::denorm_min() ) continue; // this line seems not to work like it's supposed to...
     // if(std::abs(pWeight) <= 0. ) continue;
   }
   return fWeights;
